add scss, migrate navbar
